requirements.txt
# Core Framework
gradio>=4.30.0
gradio_client>=1.0.0

# Google Cloud Services
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
vertexai>=1.38.0

# Utilities
python-dotenv>=1.0.0

# Optional: For future Neo4j integration (Phase 4)
# neo4j>=5.14.0